refactored to an auth package
