+++ app/src/main/java/com/cmput301f21t34/habittrak/auth/Auth.java
@@ -1,0 +1,4 @@
+package com.cmput301f21t34.habittrak.auth;
+
+public class Auth {
+}
