+++ app/src/main/java/com/cmput301f21t34/habittrak/auth/SignUpFragment.java
@@ -1,4 +1,4 @@
-package com.cmput301f21t34.habittrak;
+package com.cmput301f21t34.habittrak.auth;
 
 import android.content.Intent;
 import android.os.Bundle;
@@ -13,6 +13,9 @@ import android.view.View;
 import android.view.ViewGroup;
 import android.widget.EditText;
 
+import com.cmput301f21t34.habittrak.BaseActivity;
+import com.cmput301f21t34.habittrak.DatabaseManager;
+import com.cmput301f21t34.habittrak.R;
 import com.cmput301f21t34.habittrak.user.User;
 import com.google.android.material.button.MaterialButton;
 import com.google.android.material.textfield.TextInputEditText;
