+++ app/src/main/res/layout/habi_signup_fragment.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:layout_gravity="center_vertical"
-    tools:context=".SignUpFragment">
+    tools:context=".auth.SignUpFragment">
 
     <LinearLayout
         android:layout_width="match_parent"
